refactor: move errorBoundary element into separate folder
